requirements.txt
requests~=2.31.0
flask~=2.2.2
scikit-learn~=1.3.0
python-dotenv~=0.21.0